+++ src/metrics/family.rs
@@ -341,6 +341,56 @@ impl<S: Clone + std::hash::Hash + Eq, M, C: MetricConstructor<M>> Family<S, M, C
         self.metrics.write().clear()
     }
 
+    /// Returns `true` if the given label set exists within the metric family.
+    ///
+    /// ```
+    /// # use prometheus_client::metrics::counter::{Atomic, Counter};
+    /// # use prometheus_client::metrics::family::Family;
+    /// #
+    /// let family = Family::<Vec<(String, String)>, Counter>::default();
+    /// let get = vec![("method".to_owned(), "GET".to_owned())];
+    /// let post = vec![("method".to_owned(), "POST".to_owned())];
+    ///
+    /// // Create the metric with labels `method="GET"`.
+    /// family.get_or_create(&get).inc();
+    ///
+    /// assert!(family.contains(&get), "a `method=\"GET\"`-labeled metric exists");
+    /// assert!(!family.contains(&post), "a `method=\"POST\"`-labeled metric does NOT exist");
+    /// ```
+    #[cfg(any(test, feature = "test-util"))]
+    pub fn contains(&self, label_set: &S) -> bool {
+        self.metrics.read().get(label_set).is_some()
+    }
+
+    /// Returns the number of metrics in this family.
+    ///
+    /// ```
+    /// # use prometheus_client::metrics::counter::{Atomic, Counter};
+    /// # use prometheus_client::metrics::family::Family;
+    /// #
+    /// let family = Family::<Vec<(String, String)>, Counter>::default();
+    /// assert_eq!(family.len(), 0);
+    ///
+    /// // Will create the metric with label `method="GET"` on first call and
+    /// // return a reference.
+    /// family.get_or_create(&vec![("method".to_owned(), "GET".to_owned())]).inc();
+    /// assert_eq!(family.len(), 1);
+    ///
+    /// // Clear the family of all label sets.
+    /// family.clear();
+    /// assert_eq!(family.len(), 0);
+    /// ```
+    #[cfg(any(test, feature = "test-util"))]
+    pub fn len(&self) -> usize {
+        self.metrics.read().len()
+    }
+
+    /// Returns `true` if the family contains no metrics.
+    #[cfg(any(test, feature = "test-util"))]
+    pub fn is_empty(&self) -> bool {
+        self.metrics.read().is_empty()
+    }
+
     pub(crate) fn read(&self) -> RwLockReadGuard<'_, HashMap<S, M>> {
         self.metrics.read()
     }
@@ -376,6 +426,10 @@ where
 
     fn metric_type(&self) -> MetricType {
         M::TYPE
+    }
+
+    fn is_empty(&self) -> bool {
+        self.metrics.read().is_empty()
     }
 }